# apps/web/src/lib/contracts/openapi.generated.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "hubris-server",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
    "/api/chats/{conversation_id}": {
      "get": {
        "tags": ["crate::api::chats"],
        "operationId": "get_chat",
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "description": "Conversation ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Conversation detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatConversationDetail"
                }
              }
            }
          },
          "404": {
            "description": "Conversation not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/chats/{conversation_id}/interrupt": {
      "post": {
        "tags": ["crate::api::chats"],
        "operationId": "interrupt_chat",
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "description": "Conversation ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Interrupt requested"
          },
          "404": {
            "description": "Conversation not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conversation not running",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/chats/{conversation_id}/messages": {
      "post": {
        "tags": ["crate::api::chats"],
        "operationId": "send_chat_message",
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "description": "Conversation ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendChatMessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Message accepted"
          },
          "404": {
            "description": "Conversation not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Failed to start Codex runtime",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/editor-themes": {
      "get": {
        "tags": ["crate::api::editor_themes"],
        "summary": "GET /api/editor-themes — list all available editor themes.",
        "operationId": "list_editor_themes",
        "responses": {
          "200": {
            "description": "Editor theme list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EditorThemeEntry"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["crate::api::editor_themes"],
        "summary": "POST /api/editor-themes — upload a custom VS Code theme.",
        "operationId": "upload_editor_theme",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VscodeThemeJson"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Theme created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditorThemeEntry"
                }
              }
            }
          },
          "400": {
            "description": "Invalid theme JSON"
          }
        }
      }
    },
...
    "/api/editor-themes/import": {
      "post": {
        "tags": ["crate::api::editor_themes"],
        "summary": "POST /api/editor-themes/import — import a theme from an extension.",
        "operationId": "import_extension_theme",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportThemeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Theme imported",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditorThemeEntry"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "404": {
            "description": "Extension or theme not found"
          }
        }
      }
    },
    "/api/editor-themes/{id}": {
      "get": {
        "tags": ["crate::api::editor_themes"],
        "summary": "GET /api/editor-themes/{id} — full theme data.",
        "operationId": "get_editor_theme",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Theme ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Theme data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VscodeThemeJson"
                }
              }
            }
          },
          "404": {
            "description": "Theme not found"
          }
        }
      },
      "delete": {
        "tags": ["crate::api::editor_themes"],
        "summary": "DELETE /api/editor-themes/{id} — delete a custom theme.",
        "operationId": "delete_editor_theme",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Theme ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Theme deleted"
          },
          "403": {
            "description": "Cannot delete built-in theme"
          },
          "404": {
            "description": "Theme not found"
          }
        }
      }
    },
    "/api/events": {
      "get": {
        "tags": ["crate::api::events"],
        "operationId": "event_stream",
        "parameters": [
          {
            "name": "session_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Server-sent event stream"
          }
        }
      }
    },
    "/api/files": {
      "get": {
        "tags": ["crate::api::files"],
        "operationId": "list_files",
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "description": "Directory to list. Defaults to home dir.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "show_hidden",
            "in": "query",
            "description": "Whether to include dotfiles. Defaults to false.",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Directory listing",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListFilesResponse"
                }
              }
            }
          },
          "400": {
            "description": "Path is not a directory"
          },
          "403": {
            "description": "Permission denied"
          },
          "404": {
            "description": "Directory not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/keybindings": {
      "get": {
        "tags": ["crate::api::keybindings"],
        "summary": "GET /api/keybindings",
        "operationId": "get_keybindings",
        "responses": {
          "200": {
            "description": "Current keybindings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KeybindingsState"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "put": {
        "tags": ["crate::api::keybindings"],
        "summary": "PUT /api/keybindings",
        "operationId": "put_keybindings",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/KeybindingEntry"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Keybindings saved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KeybindingsState"
                }
              }
            }
          },
          "409": {
            "description": "Keybindings file is invalid on disk"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/api/tabs/reorder": {
      "put": {
        "tags": ["crate::api::tabs"],
        "operationId": "reorder_tabs",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderTabsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Tabs reordered",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TabInfo"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiErrorResponse": {
        "type": "object",
        "required": ["message"],
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
      "ChatConversationDetail": {
        "type": "object",
        "description": "Full chat detail payload used to hydrate an open chat tab.",
        "required": ["conversation", "messages"],
        "properties": {
          "conversation": {
            "$ref": "#/components/schemas/ChatConversationSummary"
          },
          "latestRun": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ChatRun"
              }
            ]
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChatMessage"
            }
          }
        }
      },
...
      "EditorThemeEntry": {
        "type": "object",
        "required": ["id", "name", "type", "builtin"],
        "properties": {
          "builtin": {
            "type": "boolean"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        }
      },
...
      "SendChatMessageRequest": {
        "type": "object",
        "required": ["text"],
        "properties": {
          "text": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}